refactor contactForm file structure Rename contactForm -> ContactForm Add barrel file CSS file is together with the jsx file Update imports
